changed opcode type from u16 to u8

+++ src/cpu/common_defs/mod.rs
@@ -3,12 +3,13 @@ pub mod address_mode;
 pub mod opcode_class;
 
 //private mods
+use std::clone::Clone;
 
 // hoisted interfaces
 
 #[derive(Default)]
 pub struct OpcodeDebugInfo {
-    pub opcode: u16,
+    pub opcode: u8,
     pub name: String,
     pub address_mode_name: String,
     pub notes: String,
@@ -16,7 +17,7 @@ pub struct OpcodeDebugInfo {
 
 #[derive(Default)]
 pub struct OpcodeExecInfo {
-    pub opcode: u16,
+    pub opcode: u8,
     pub opcode_class:opcode_class::OpcodeClass,
     pub address_mode:address_mode::AddressMode,
     pub len: u8,
